[FEAT] 매칭 기록 조회 api

See merge request s08-bigdata-recom-sub2/S08P22D204!115

+++ src/main/java/com/d204/rumeet/record/model/dto/RaceInfoSummaryDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class RaceInfoSummaryDto {
     long total_km;
     long total_time;
-    long avg_pace;
+    int avg_pace;
 }
